+++ hydra-node/src/Hydra/HeadLogic.hs
@@ -39,7 +39,7 @@ import Hydra.Chain (
   rollbackHistory,
   setLastKnown,
  )
-import Hydra.Chain.ChainState (ChainSlot, IsChainState (..), chainStateSlot)
+import Hydra.Chain.ChainState (ChainSlot (..), IsChainState (..), chainStateSlot)
 import Hydra.HeadLogic.Error (
   LogicError (..),
   RequirementFailure (..),
@@ -1330,27 +1330,32 @@ onClosedChainFanoutTx closedState newChainState fanoutUTxO =
 -- | Detect our view of the chain going out of sync and issue a 'NodeUnsynced'
 -- event when this is the case.
 handleOutOfSync ::
+  IsChainState tx =>
   Environment ->
   -- | Current system time
   UTCTime ->
-  -- | Chain time
+  -- | Latest Chain point observed
+  ChainPointType tx ->
+  -- | Latest Chain point time representation observed
   UTCTime ->
   SyncedStatus ->
   Outcome tx
-handleOutOfSync Environment{unsyncedPeriod} now chainTime syncStatus
+handleOutOfSync Environment{unsyncedPeriod} now chainPoint chainTime syncStatus
   -- We consider the node out of sync when:
   -- the last observed chainTime plus the delta allowed by the unsyncedPeriod
   -- falls behind the current system time.
   | chainTime `plus` unsyncedPeriodToNominalDiffTime unsyncedPeriod < now =
       case syncStatus of
-        InSync -> newState NodeUnsynced
+        InSync -> newState (NodeUnsynced{chainSlot, chainTime, drift})
         CatchingUp -> noop
   | otherwise =
       case syncStatus of
         InSync -> noop
-        CatchingUp -> newState NodeSynced
+        CatchingUp -> newState (NodeSynced{chainSlot, chainTime, drift})
  where
   plus = flip addUTCTime
+  chainSlot = chainPointSlot chainPoint
+  drift = now `diffUTCTime` chainTime
 
 -- | Handles inputs and converts them into 'StateChanged' events along with
 -- 'Effect's, in case it is processed successfully. Later, the Node will
@@ -1368,8 +1373,8 @@ update ::
   Outcome tx
 update env ledger now nodeState ev =
   case nodeState of
-    NodeCatchingUp{headState, pendingDeposits, currentSlot} ->
-      updateUnsyncedHead env ledger now currentSlot pendingDeposits headState ev (syncedStatus nodeState)
+    NodeCatchingUp{headState, pendingDeposits, currentSlot, currentChainTime} ->
+      updateUnsyncedHead env ledger now currentSlot currentChainTime pendingDeposits headState ev (syncedStatus nodeState)
     NodeInSync{headState, pendingDeposits, currentSlot} ->
       updateSyncedHead env ledger now currentSlot pendingDeposits headState ev (syncedStatus nodeState)
 
@@ -1379,7 +1384,10 @@ updateUnsyncedHead ::
   Ledger tx ->
   -- | Current system time.
   UTCTime ->
+  -- | Node's current chain slot
   ChainSlot ->
+  -- | Latest Chain point time representation observed
+  UTCTime ->
   PendingDeposits tx ->
   -- | Current HeadState to validate the command against.
   HeadState tx ->
@@ -1387,14 +1395,16 @@ updateUnsyncedHead ::
   Input tx ->
   SyncedStatus ->
   Outcome tx
-updateUnsyncedHead env ledger now currentSlot pendingDeposits st ev syncStatus =
+updateUnsyncedHead env ledger now currentSlot currentChainTime pendingDeposits st ev syncStatus =
   case ev of
     ChainInput{} ->
       handleChainInput env ledger now currentSlot pendingDeposits st ev syncStatus
     ClientInput{clientInput} ->
-      cause . ClientEffect $ ServerOutput.RejectedInput clientInput "chain out of sync"
+      cause . ClientEffect $ ServerOutput.RejectedInput clientInput $ "chain out of sync, drift: " <> show drift
     NetworkInput{} ->
       wait WaitOnNodeInSync{currentSlot}
+ where
+  drift = now `diffUTCTime` currentChainTime
 
 updateSyncedHead ::
   IsChainState tx =>
@@ -1402,6 +1412,7 @@ updateSyncedHead ::
   Ledger tx ->
   -- | Current system time.
   UTCTime ->
+  -- | Node's current chain slot
   ChainSlot ->
   PendingDeposits tx ->
   -- | Current HeadState to validate the command against.
@@ -1427,6 +1438,7 @@ handleChainInput ::
   Ledger tx ->
   -- | Current system time.
   UTCTime ->
+  -- | Node's current chain slot
   ChainSlot ->
   PendingDeposits tx ->
   -- | Current HeadState to validate the command against.
@@ -1464,7 +1476,7 @@ handleChainInput env _ledger now _currentSlot pendingDeposits st ev syncStatus =
     -- time did not advance in an open head anymore. This is a hint that we
     -- should compose event handling better.
     newState TickObserved{chainPoint}
-      <> handleOutOfSync env now chainTime syncStatus
+      <> handleOutOfSync env now chainPoint chainTime syncStatus
       <> onChainTick env pendingDeposits chainTime
       <> onOpenChainTick env chainTime (depositsForHead ourHeadId pendingDeposits) openState
   (Open openState@OpenState{headId = ourHeadId}, ChainInput Observation{observedTx = OnIncrementTx{headId, newVersion, depositTxId}, newChainState})
@@ -1487,7 +1499,7 @@ handleChainInput env _ledger now _currentSlot pendingDeposits st ev syncStatus =
   (Closed ClosedState{contestationDeadline, readyToFanoutSent, headId}, ChainInput Tick{chainTime, chainPoint})
     | chainTime > contestationDeadline && not readyToFanoutSent ->
         newState TickObserved{chainPoint}
-          <> handleOutOfSync env now chainTime syncStatus
+          <> handleOutOfSync env now chainPoint chainTime syncStatus
           <> onChainTick env pendingDeposits chainTime
           <> newState HeadIsReadyToFanout{headId}
   (Closed closedState@ClosedState{headId = ourHeadId}, ChainInput Observation{observedTx = OnFanoutTx{headId, fanoutUTxO}, newChainState})
@@ -1503,10 +1515,10 @@ handleChainInput env _ledger now _currentSlot pendingDeposits st ev syncStatus =
   -- General
   (_, ChainInput Rollback{rolledBackChainState, chainTime}) ->
     newState ChainRolledBack{chainState = rolledBackChainState}
-      <> handleOutOfSync env now chainTime syncStatus
+      <> handleOutOfSync env now (chainStatePoint rolledBackChainState) chainTime syncStatus
   (_, ChainInput Tick{chainTime, chainPoint}) ->
     newState TickObserved{chainPoint}
-      <> handleOutOfSync env now chainTime syncStatus
+      <> handleOutOfSync env now chainPoint chainTime syncStatus
       <> onChainTick env pendingDeposits chainTime
   (_, ChainInput PostTxError{postChainTx, postTxError}) ->
     cause . ClientEffect $ ServerOutput.PostTxOnChainFailed{postChainTx, postTxError}
@@ -1519,6 +1531,7 @@ handleNetworkInput ::
   Ledger tx ->
   -- | Current system time.
   UTCTime ->
+  -- | Node's current chain slot
   ChainSlot ->
   PendingDeposits tx ->
   -- | Current NodeState to validate the command against.
@@ -1562,6 +1575,7 @@ handleClientInput ::
   Ledger tx ->
   -- | Current system time.
   UTCTime ->
+  -- | Node's current chain slot
   ChainSlot ->
   PendingDeposits tx ->
   -- | Current NodeState to validate the command against.
@@ -1666,10 +1680,10 @@ aggregateNodeState nodeState sc =
           nodeState{headState = st, currentSlot = chainPointSlot chainPoint}
         ChainRolledBack{chainState} ->
           nodeState{headState = st, currentSlot = chainStateSlot chainState}
-        NodeUnsynced ->
-          NodeCatchingUp{headState = st, pendingDeposits = currentPendingDeposits, currentSlot = nodeState.currentSlot}
-        NodeSynced ->
-          NodeInSync{headState = st, pendingDeposits = currentPendingDeposits, currentSlot = nodeState.currentSlot}
+        NodeUnsynced{chainSlot, chainTime} ->
+          NodeCatchingUp{headState = st, pendingDeposits = currentPendingDeposits, currentSlot = chainSlot, currentChainTime = chainTime}
+        NodeSynced{chainSlot, chainTime} ->
+          NodeInSync{headState = st, pendingDeposits = currentPendingDeposits, currentSlot = chainSlot, currentChainTime = chainTime}
         _ ->
           nodeState{headState = st}
 
@@ -1956,8 +1970,8 @@ aggregate st = \case
       Open ost{coordinatedHeadState = coordState{allTxs = foldr Map.delete allTransactions [txId transaction]}}
     _otherState -> st
   Checkpoint nodeState -> headState nodeState
-  NodeSynced -> st
-  NodeUnsynced -> st
+  NodeSynced{} -> st
+  NodeUnsynced{} -> st
 
 aggregateState ::
   IsChainState tx =>
@@ -2012,5 +2026,5 @@ aggregateChainStateHistory history = \case
   LocalStateCleared{} -> history
   -- FIXME: This makes chain sync starting after rollbacks past the chain state impossible
   Checkpoint nodeState -> initHistory $ getChainState nodeState.headState
-  NodeUnsynced -> history
-  NodeSynced -> history
+  NodeUnsynced{} -> history
+  NodeSynced{} -> history