+++ hydra-node/testlib/Test/Hydra/HeadLogic/Outcome.hs
@@ -62,8 +62,8 @@ genStateChanged env =
     , HeadIsReadyToFanout <$> arbitrary
     , HeadFannedOut <$> arbitrary <*> arbitrary <*> arbitrary
     , LocalStateCleared <$> arbitrary <*> arbitrary
-    , pure NodeUnsynced
-    , pure NodeSynced
+    , NodeUnsynced <$> arbitrary <*> arbitrary <*> arbitrary
+    , NodeSynced <$> arbitrary <*> arbitrary <*> arbitrary
     ]
  where
   Environment{party} = env